refactor: :recycle: Move ~/utils/requests/models to ~/utils/models.
