+++ utils/models/types/strings.ts
