+++ utils/requests/index.ts
@@ -2,4 +2,3 @@ export * from "./login";
 export * from "./register";
 export * from "./account/activate";
 export * from "./articles";
-export * from "./models";
